+++ backend/src/main/java/com/project/backend/domain/book/util/BookUtil.java
@@ -1,4 +1,4 @@
-package com.project.backend.domain.book.service;
+package com.project.backend.domain.book.util;
 
 /**
  * --도서에 쓰이는 유틸관련 클래스--
@@ -12,7 +12,7 @@ public class BookUtil {
      * 카카오 API에서 제공하는 ISBN은 "10자리ISBN 13자리ISBN" 형식이므로
      * 13자리 ISBN만 추출하여 반환한다.
      *
-     * @param String isbn 
+     * @param --String isbn --
      * @return String 정제된 13자리 ISBN
      * @author -- 정재익 --
      * @since -- 2월 5일 --
